feat: update GitHub Actions workflow for publishing
Prior to this commit, the GitHub Actions workflow only triggered on version tags and did
not include a separate publish job. This limited the flexibility of the CI/CD pipeline.

In this commit, the workflow is updated to trigger on all branch pushes and includes a
new publish job that runs on the main branch when a tag is pushed. This enhances the
CI/CD pipeline allows dev branches to trigger builds, while preventing such releases
from being published, supporting pre-PR force push workflows.