build: add hatch-vcs dependency to project
Prior to this commit, the project did not include the `hatch-vcs` dependency, which is
necessary for version control system integration with Hatch.

In this commit, the `hatch-vcs` dependency is added to the `pixi.lock` and
`pyproject.toml` files, ensuring that the project can utilize version control features
provided by Hatch.