ci(github): enhance actions workflow
Prior to this commit, the GitHub Actions workflow only ran tests and built the package
without performing linting or managing build artifacts. This limited the CI/CD
pipeline's capabilities in terms of code quality checks and artifact handling.

In this commit, update the workflow to include a linting step before running tests, and
to upload build artifacts after the build step. Additionally, update the publish job to
download these artifacts before uploading them as release assets. This avoids
unnecessarily re-building the package in the `publish` step.