ci(github): split build and release into separate workflows
Prior to this commit, the build and release jobs were in the same workflow. This
prevented different triggers to be applied to the build and release stages. For
example, it was not possible to have _every_ pushed commit trigger a build, but
only have tagged commits trigger a release.

In this commit, the release stage is split out as a separate workflow. This
allows the build and release stages to have different triggers. The release
workflow is triggered by a tag being pushed AND the build workflow completing
successfully.